# openapi.yaml
openapi: 3.0.3
info:
  title: Express Stripe Auth API
  version: 0.1.0
  description: Squelette API – health, auth, Stripe webhook
servers:
  - url: http://localhost:3000
    description: Local
paths:
...
  /auth/register:
    post:
      summary: Register (not implemented)
      operationId: register
      tags: [Auth]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        "501":
          description: Not implemented
  /auth/login:
    post:
      summary: Login (not implemented)
      operationId: login
      tags: [Auth]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "501":
          description: Not implemented
...